introduce preload cli
