+++ vllm/config/load.py
@@ -39,9 +39,8 @@ class LoadConfig:
       InstantTensor, which enables distributed loading with pipelined prefetching
       and fast direct I/O.
     - "ipc_cache" will map post-quantized weights from a local weight cache
-      daemon via CUDA IPC for fast engine restarts. See
-      `vllm/model_executor/model_loader/weight_cache/daemon.py` for how to
-      launch the daemon.
+      daemon via CUDA IPC for fast engine restarts. Launch the daemon with
+      `vllm preload` first.
     - "npcache" will load the weights in pytorch format and store a numpy cache
       to speed up the loading.
     - "dummy" will initialize the weights with random values, which is mainly
