+++ vllm/entrypoints/cli/main.py
@@ -44,6 +44,7 @@ def main():
     import vllm.entrypoints.cli.collect_env
     import vllm.entrypoints.cli.launch
     import vllm.entrypoints.cli.openai
+    import vllm.entrypoints.cli.preload
     import vllm.entrypoints.cli.run_batch
     import vllm.entrypoints.cli.serve
     from vllm.entrypoints.serve.utils.api_utils import (
@@ -58,6 +59,7 @@ def main():
         vllm.entrypoints.cli.launch,
         vllm.entrypoints.cli.benchmark.main,
         vllm.entrypoints.cli.collect_env,
+        vllm.entrypoints.cli.preload,
         vllm.entrypoints.cli.run_batch,
     ]
